requirements.txt
# FungiCloud - Servidor Cloud
# Python 3.9+

# Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Base de datos
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Autenticación
PyJWT==2.8.0
bcrypt==4.1.2

# Stripe
stripe==7.8.0

# Utilidades
python-dotenv==1.0.0

# Servidor WSGI para producción
gunicorn==21.2.0
